// src/serverApi/v3/models/room-item-response-allowed-operations.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Schulcloud-Verbund-Software Server API
 * This is v3 of Schulcloud-Verbund-Software Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface RoomItemResponseAllowedOperations
 */
export interface RoomItemResponseAllowedOperations {
    /**
     * 
     * @type {boolean}
     * @memberof RoomItemResponseAllowedOperations
     */
    accessRoom: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof RoomItemResponseAllowedOperations
     */
    accessRoomBoards: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof RoomItemResponseAllowedOperations
     */
    addAllStudents: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof RoomItemResponseAllowedOperations
     */
    addExternalPersonByEmail: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof RoomItemResponseAllowedOperations
     */
    addMembers: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof RoomItemResponseAllowedOperations
     */
    arrangeRooms: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof RoomItemResponseAllowedOperations
     */
    changeRolesOfMembers: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof RoomItemResponseAllowedOperations
     */
    copyRoom: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof RoomItemResponseAllowedOperations
     */
    createRoom: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof RoomItemResponseAllowedOperations
     */
    deleteRoom: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof RoomItemResponseAllowedOperations
     */
    editContent: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof RoomItemResponseAllowedOperations
     */
    getRoomMembers: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof RoomItemResponseAllowedOperations
     */
    getRoomMembersRedacted: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof RoomItemResponseAllowedOperations
     */
    leaveRoom: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof RoomItemResponseAllowedOperations
     */
    shareRoom: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof RoomItemResponseAllowedOperations
     */
    updateRoom: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof RoomItemResponseAllowedOperations
     */
    createRoomInvitationLinks: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof RoomItemResponseAllowedOperations
     */
    listRoomInvitationLinks: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof RoomItemResponseAllowedOperations
     */
    updateRoomInvitationLinks: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof RoomItemResponseAllowedOperations
     */
    deleteRoomInvitationLinks: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof RoomItemResponseAllowedOperations
     */
    viewContent: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof RoomItemResponseAllowedOperations
     */
    viewDraftContent: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof RoomItemResponseAllowedOperations
     */
    viewMemberlist: boolean;
}